MSAgent/requirements.txt
agent-framework>=1.0.0b251028
pyodbc>=5.0.0
python-dotenv>=1.0.0
